requirements.txt
fastapi
uvicorn
firebase-admin
datetime
google-cloud-firestore
google-generativeai
python-dotenv